Finally solved!!!
* https://stackoverflow.com/a/63848865
* https://docs.rs/diesel/1.4.7/diesel/sql_types/struct.Integer.html

+++ src/models.rs
@@ -9,12 +9,12 @@ use std::time::SystemTime;
 
 #[derive(Queryable)]
 pub struct Post {
-    pub id: i64,
+    pub id: i32,
     pub title: String,
     pub body: String,
     pub published: bool,
     pub publish_at: Option<SystemTime>,
-    pub visit_count: Option<i64>,
+    pub visit_count: Option<i32>,
 }
 
 #[derive(Insertable)]
@@ -23,5 +23,5 @@ pub struct NewPost<'a> {
     pub title: &'a str,
     pub body: &'a str,
     pub publish_at: Option<SystemTime>,
-    pub visit_count: Option<i64>,
+    pub visit_count: Option<i32>,
 }
